Moved remaining numbers and messages to constant class and added interface file IRender.cs for ConsoleRender.cs.
