+++ GameOfLife/Interfaces/IRenderer.cs
@@ -1,0 +1,15 @@
+using GameOfLife.Models;
+
+namespace GameOfLife.Interfaces
+{
+    /// <summary>
+    /// Interface for game renderers, allowing structured rendering implementations
+    /// </summary>
+    public interface IRenderer
+    {
+        void Render(Grid grid);
+        (int rows, int columns) GetGridSize();
+        void DisplayInvalidInputMessage();
+        void DisplayGameControls();
+    }
+}
